# contracts/openapi/openapi-v1.json
{
  "components": {
    "schemas": {
      "ApiErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayload"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProfileData": {
        "properties": {
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "phone_e164": {
            "title": "Phone E164",
            "type": "string"
          },
          "phone_number": {
            "title": "Phone Number",
            "type": "string"
          },
          "phone_prefix": {
            "title": "Phone Prefix",
            "type": "string"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "display_name",
          "country_code",
          "phone_prefix",
          "phone_number",
          "phone_e164"
        ],
        "title": "ProfileData",
        "type": "object"
      },
      "ProfileEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProfileData"
          },
          "ok": {
            "const": true,
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "ProfileEnvelope",
        "type": "object"
      },
      "ProfileUpdateRequest": {
        "properties": {
          "country_code": {
            "anyOf": [
              {
                "maxLength": 8,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "phone_number": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 6,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "phone_prefix": {
            "anyOf": [
              {
                "maxLength": 8,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Prefix"
          },
          "photo_url": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "title": "ProfileUpdateRequest",
        "type": "object"
      },
      "TripCreateRequest": {
        "properties": {
          "departure_at": {
            "maxLength": 64,
            "minLength": 8,
            "title": "Departure At",
            "type": "string"
          },
          "direction": {
            "maxLength": 20,
            "minLength": 3,
            "title": "Direction",
            "type": "string"
          },
          "origin_label": {
            "maxLength": 160,
            "minLength": 3,
            "title": "Origin Label",
            "type": "string"
          },
          "total_seats": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Total Seats",
            "type": "integer"
          },
          "vehicle_id": {
            "minimum": 1.0,
            "title": "Vehicle Id",
            "type": "integer"
          }
        },
        "required": [
          "direction",
          "origin_label",
          "departure_at",
          "total_seats",
          "vehicle_id"
        ],
        "title": "TripCreateRequest",
        "type": "object"
      },
...
      "TripEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripData"
          },
          "ok": {
            "const": true,
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "TripEnvelope",
        "type": "object"
      },
...
      "TripsListEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripsListData"
          },
          "ok": {
            "const": true,
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "TripsListEnvelope",
        "type": "object"
      },
...
      "VehicleCreateRequest": {
        "properties": {
          "brand": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Brand",
            "type": "string"
          },
          "color": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Color",
            "type": "string"
          },
          "plate": {
            "maxLength": 12,
            "minLength": 5,
            "title": "Plate",
            "type": "string"
          },
          "reference": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Reference",
            "type": "string"
          }
        },
        "required": [
          "brand",
          "reference",
          "color",
          "plate"
        ],
        "title": "VehicleCreateRequest",
        "type": "object"
      },
...
      "VehicleEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VehicleData"
          },
          "ok": {
            "const": true,
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "VehicleEnvelope",
        "type": "object"
      },
...
      "VehiclesListEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VehiclesListData"
          },
          "ok": {
            "const": true,
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "VehiclesListEnvelope",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Transporte Vecinal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "get_my_profile_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get My Profile",
        "tags": [
          "profile"
        ]
      },
      "patch": {
        "operationId": "update_my_profile_api_v1_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Update My Profile",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/trips": {
      "get": {
        "operationId": "list_trips_api_v1_trips_get",
        "parameters": [
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Direction"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripsListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Trips",
        "tags": [
          "trips"
        ]
      },
      "post": {
        "operationId": "create_trip_api_v1_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Trip",
        "tags": [
          "trips"
        ]
      }
    },
...
    "/api/v1/trips/{trip_id}/cancel": {
      "post": {
        "operationId": "cancel_trip_api_v1_trips__trip_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "title": "Trip Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/finalize": {
      "post": {
        "operationId": "finalize_trip_api_v1_trips__trip_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "title": "Trip Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/v1/vehicles": {
      "get": {
        "operationId": "list_vehicles_api_v1_vehicles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiclesListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Vehicles",
        "tags": [
          "vehicles"
        ]
      },
      "post": {
        "operationId": "create_vehicle_api_v1_vehicles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Vehicle",
        "tags": [
          "vehicles"
        ]
      }
    },
    "/api/v1/vehicles/{vehicle_id}": {
      "delete": {
        "operationId": "deactivate_vehicle_api_v1_vehicles__vehicle_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "title": "Vehicle Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deactivate Vehicle",
        "tags": [
          "vehicles"
        ]
      },
      "patch": {
        "operationId": "update_vehicle_api_v1_vehicles__vehicle_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "vehicle_id",
            "required": true,
            "schema": {
              "title": "Vehicle Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Vehicle",
        "tags": [
          "vehicles"
        ]
      }
    },
...
  }
}